requirements.txt
APScheduler==3.9.1
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
cssselect2==0.6.0
defusedxml==0.7.1
Django==3.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
djangorestframework-xml==2.0.0
drf-spectacular==0.21.2
et-xmlfile==1.1.0
future==0.18.2
html5lib==1.1
idna==3.3
importlib-resources==5.4.0
inflection==0.5.1
jsonschema==4.4.0
lxml==4.9.0
mssql-django==1.1.2
numpy==1.22.3
openpyxl==3.0.9
oscrypto==1.3.0
pandas==1.3.5
Pillow==9.1.1
pycparser==2.21
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyJWT==2.3.0
PyMySQL==1.0.2
pyodbc==4.0.32
PyPDF3==1.0.6
pyrsistent==0.18.1
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.10
requests==2.28.1
roman==3.3
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
xhtml2pdf==0.2.8
XlsxWriter==3.0.3
zipp==3.7.0
